Fixed precision issues issues

+++ Plugins/DayNightCycle/Source/DayNightCycle/Private/Clock/GameClockSubsystem.cpp
@@ -22,7 +22,7 @@ const TScriptInterface<IGameClock> & UGameClockSubsystem::GetGameClock() const {
 
 bool UGameClockSubsystem::IsDay() const {
     auto CurrentTime = IGameClock::Execute_GetCurrentTime(GameClock.GetObject());
-    float CurrentHour = CurrentTime.GetTimeOfDay().GetTotalHours();
+    float CurrentHour = static_cast<float>(CurrentTime.GetTimeOfDay().GetTotalHours());
     return GetDefault<UDayNightCycleSettings>()->DayRange.Contains(CurrentHour);
 }
 
@@ -32,7 +32,7 @@ float UGameClockSubsystem::GetDayCoefficient() const {
     }
     
     auto CurrentTime = IGameClock::Execute_GetCurrentTime(GameClock.GetObject());
-    float CurrentHour = CurrentTime.GetTimeOfDay().GetTotalHours();
+    float CurrentHour = static_cast<float>(CurrentTime.GetTimeOfDay().GetTotalHours());
     auto &DayRange = GetDefault<UDayNightCycleSettings>()->DayRange;
     return (CurrentHour - DayRange.GetLowerBoundValue()) / (DayRange.GetUpperBoundValue() - DayRange.GetLowerBoundValue());
 }
